Дюкин Петр 5.3

+++ 05-java-oop-principles/task03/src/com/example/task03/Seconds.java
@@ -23,11 +23,11 @@ public class Seconds implements TimeUnit {
 
     @Override
     public long toMinutes() {
-        return Math.round(amount / 60);
+        return Math.round(amount / 60.0);
     }
 
     @Override
     public long toHours() {
-        return Math.round(amount / (60 * 60));
+        return Math.round(amount /  (double) (60 * 60));
     }
 }
